Move submit translation internals into a separate module
