+++ wagtail_localize/views/submit_translations.py
@@ -1,10 +1,7 @@
-from collections import defaultdict
-
 from django import forms
-from django.conf import settings
 from django.contrib import messages
 from django.contrib.admin.utils import unquote
-from django.core.exceptions import PermissionDenied, ValidationError
+from django.core.exceptions import PermissionDenied
 from django.db import transaction
 from django.http import Http404
 from django.shortcuts import get_object_or_404, redirect
@@ -18,7 +15,7 @@ from wagtail.core.models import Locale, Page, TranslatableMixin
 from wagtail.snippets.views.snippets import get_snippet_model_from_url_params
 
 from wagtail_localize.components import TranslationComponentManager
-from wagtail_localize.models import Translation, TranslationSource
+from wagtail_localize.operations import translate_object, translate_page_subtree
 
 
 class SubmitTranslationForm(forms.Form):
@@ -69,76 +66,6 @@ class SubmitTranslationForm(forms.Form):
         # anyway and it gets cached so it'll only have one query in the end.
         if len(self.fields["locales"].queryset) < 2:
             self.fields["select_all"].widget = forms.HiddenInput()
-
-
-class TranslationCreator:
-    """
-    A class that provides a create_translations method.
-
-    Call create_translations for each object you want to translate and this will submit
-    that object and any dependencies as well.
-
-    This class will track the objects that have already submitted so an object doesn't
-    get submitted twice.
-    """
-
-    def __init__(self, user, target_locales):
-        self.user = user
-        self.target_locales = target_locales
-        self.seen_objects = set()
-        self.mappings = defaultdict(list)
-
-    def create_translations(self, instance, include_related_objects=True):
-        if isinstance(instance, Page):
-            instance = instance.specific
-
-        if instance.translation_key in self.seen_objects:
-            return
-        self.seen_objects.add(instance.translation_key)
-
-        source, created = TranslationSource.get_or_create_from_instance(instance)
-
-        # Add related objects
-        # Must be before translation records or those translation records won't be able to create
-        # the objects because the dependencies haven't been created
-        if include_related_objects:
-            for related_object_segment in source.relatedobjectsegment_set.all():
-                related_instance = related_object_segment.object.get_instance(
-                    instance.locale
-                )
-
-                # Limit to one level of related objects, since this could potentially pull in a lot of stuff
-                self.create_translations(
-                    related_instance, include_related_objects=False
-                )
-
-        # Support disabling the out of the box translation mode.
-        # The value set on the model takes precendence over the global setting.
-        if hasattr(instance, "localize_default_translation_mode"):
-            translation_mode = instance.localize_default_translation_mode
-        else:
-            translation_mode = getattr(
-                settings, "WAGTAIL_LOCALIZE_DEFAULT_TRANSLATION_MODE", "synced"
-            )
-        translation_enabled = translation_mode == "synced"
-
-        # Set up translation records
-        for target_locale in self.target_locales:
-            # Create translation if it doesn't exist yet, re-enable if translation was disabled
-            # Note that the form won't show this locale as an option if the translation existed
-            # in this langauge, so this shouldn't overwrite any unmanaged translations.
-            translation, created = Translation.objects.update_or_create(
-                source=source,
-                target_locale=target_locale,
-                defaults={"enabled": translation_enabled},
-            )
-
-            self.mappings[source].append(translation)
-
-            try:
-                translation.save_target(user=self.user)
-            except ValidationError:
-                pass
 
 
 class SubmitTranslationView(SingleObjectMixin, TemplateView):
@@ -194,21 +121,21 @@ class SubmitTranslationView(SingleObjectMixin, TemplateView):
 
     @transaction.atomic
     def form_valid(self, form):
-        translator = TranslationCreator(self.request.user, form.cleaned_data["locales"])
-        translator.create_translations(self.object)
+        translate_object(
+            self.object,
+            form.cleaned_data["locales"],
+            self.components,
+            self.request.user,
+        )
 
         # Now add the sub tree (if the obj is a page)
-        if isinstance(self.object, Page):
-            if form.cleaned_data["include_subtree"]:
-
-                def _walk(current_page):
-                    for child_page in current_page.get_children():
-                        translator.create_translations(child_page)
-
-                        if child_page.numchild:
-                            _walk(child_page)
-
-                _walk(self.object)
+        if isinstance(self.object, Page) and form.cleaned_data["include_subtree"]:
+            translate_page_subtree(
+                self.object,
+                form.cleaned_data["locales"],
+                self.components,
+                self.request.user,
+            )
 
         single_translated_object = None
         if len(form.cleaned_data["locales"]) == 1:
@@ -220,8 +147,6 @@ class SubmitTranslationView(SingleObjectMixin, TemplateView):
         else:
             # Note: always plural
             locales = _("{} locales").format(len(form.cleaned_data["locales"]))
-
-        self.components.save(self.object, sources_and_translations=translator.mappings)
 
         # TODO: Button that links to page in translations report when we have it
         messages.success(self.request, self.get_success_message(locales))
